refactor(db): extract status retention repository
